Move CompetitionsLoaderPresentationAdapter to separate file
