+++ PeteBJJCompFeed/PeteBJJCompFeediOS/Feed UI/Composers/CompetitionsUIComposer.swift
@@ -62,26 +62,3 @@ private final class CompetitionsViewAdapter: CompetitionsView {
         }
     }
 }
-
-private final class CompetitionsLoaderPresentationAdapter: CompetitionsViewControllerDelegate {
-    private let feedLoader: CompetitionsLoader
-    var presenter: CompetitionsPresenter?
-
-    init(feedLoader: CompetitionsLoader) {
-        self.feedLoader = feedLoader
-    }
-
-    func didRequestFeedRefresh() {
-        presenter?.didStartLoadingFeed()
-
-        feedLoader.load { [weak self] result in
-            switch result {
-            case let .success(feed):
-                self?.presenter?.didFinishLoadingFeed(with: feed)
-
-            case let .failure(error):
-                self?.presenter?.didFinishLoadingFeed(with: error)
-            }
-        }
-    }
-}
